ShellCheck and other fixes to provisioning scripts

+++ provisioning/shila-dev/bootstrap-host.sh
@@ -1,10 +1,11 @@
 #!/usr/bin/env bash
 
 MY_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-source ${MY_DIR}/scripts-conf.sh
+# shellcheck source=./scripts.conf
+source "${MY_DIR}/scripts-conf.sh"
 
 export DEBIAN_FRONTEND=noninteractive
-timedatectl set-timezone ${TIMEZONE}
+timedatectl set-timezone "${TIMEZONE}"
 
 
 ################################################################################
@@ -13,7 +14,7 @@ timedatectl set-timezone ${TIMEZONE}
 
 apt-get update
 
-# Some common dependencies
+# Some common dependencies.
 apt-get -y install unzip
 
 # MySQL
@@ -22,7 +23,9 @@ debconf-set-selections <<< "mysql-server mysql-server/root_password_again passwo
 apt-get -y install mysql-server mysql-client
 
 # PHP 7.2
-apt-get -y install php7.2-bcmath php7.2-bz2 php7.2-cli php7.2-curl php7.2-dev php7.2-fpm php7.2-gd php7.2-geoip php7.2-imagick php7.2-intl php7.2-mbstring php7.2-mysql php7.2-xml php7.2-zip
+apt-get -y install php7.2-bcmath php7.2-bz2 php7.2-cli php7.2-curl php7.2-dev \
+  php7.2-fpm php7.2-gd php7.2-geoip php7.2-imagick php7.2-intl php7.2-mbstring \
+  php7.2-mysql php7.2-xml php7.2-zip
 
 # Nginx
 apt-get -y install nginx
@@ -32,7 +35,7 @@ curl -sS https://getcomposer.org/installer | php
 mv composer.phar /usr/local/bin/composer
 
 # Drush 8
-wget -nv https://github.com/drush-ops/drush/releases/download/8.1.17/drush.phar
+wget -nv https://github.com/drush-ops/drush/releases/download/8.3.0/drush.phar
 chmod +x drush.phar
 mv drush.phar /usr/local/bin/drush
 
@@ -47,16 +50,16 @@ chmod +x /usr/local/bin/drupal
 ################################################################################
 
 # Nginx
-ln -sf ${CONF_ROOT}/etc/nginx/nginx.conf /etc/nginx/nginx.conf
-ln -sf ${CONF_ROOT}/etc/nginx/nginx-drupal.conf /etc/nginx/nginx-drupal.conf
-ln -sf ${CONF_ROOT}/etc/nginx/nginx-drupal6.conf /etc/nginx/nginx-drupal6.conf
-rm -rf /etc/nginx/sites-available
-ln -sf ${CONF_ROOT}/etc/nginx/sites-available /etc/nginx/sites-available
-rm /etc/nginx/sites-enabled/default
+ln -sf "${CONF_ROOT}/etc/nginx/nginx.conf" /etc/nginx/nginx.conf
+ln -sf "${CONF_ROOT}/etc/nginx/nginx-drupal.conf" /etc/nginx/nginx-drupal.conf
+ln -sf "${CONF_ROOT}/etc/nginx/nginx-drupal6.conf" /etc/nginx/nginx-drupal6.conf
+if [ -d /etc/nginx/sites-available ]; then rm -rf /etc/nginx/sites-available; fi
+ln -sf "${CONF_ROOT}/etc/nginx/sites-available" /etc/nginx/sites-available
+if [ -f /etc/nginx/sites-enabled/default ]; then rm /etc/nginx/sites-enabled/default; fi
 ln -sf /etc/nginx/sites-available/local/www.shila.test /etc/nginx/sites-enabled/www.shila.test
 ln -sf /etc/nginx/sites-available/local/pl.shila.test /etc/nginx/sites-enabled/pl.shila.test
 
-# Prepare instance directories
-mkdir -p ${INSTANCE_DIR}
-test ${OWNER_USER} != ${VAGRANT_USER} && (chown -R ${OWNER_USER}:${OWNER_USER} ${INSTANCES_ROOT})
-ln -sf ${INSTANCE_DIR} ${SHILA_ROOT}
+# Prepare instance directories.
+mkdir -p "${INSTANCE_DIR}"
+test "${OWNER_USER}" != "${VAGRANT_USER}" && ( chown -R "${OWNER_USER}":"${OWNER_USER}" "${INSTANCES_ROOT}" )
+if [ ! -f "${SHILA_ROOT}" ]; then ln -sf "${INSTANCE_DIR}" "${SHILA_ROOT}"; fi
